Move tchap unit-tests into single folder, and run jest on that folder only
